# docs/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API for the Noterian note-taking application.\nAuthentication is performed via a JWT-cookie set by /signup and /signin.\nState-changing endpoints additionally require an X-CSRF-Token header (obtain it from /csrf-token).",
        "title": "WHITECROWSOFT API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/",
    "paths": {
...
        "/logout": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Сбрасывает JWT-cookie текущей сессии.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Выход пользователя",
                "responses": {
                    "204": {
                        "description": "Успешный выход, тело ответа отсутствует"
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/signin": {
            "post": {
                "description": "Проверяет учётные данные и устанавливает JWT-cookie сессии.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Аутентификация пользователя",
                "parameters": [
                    {
                        "description": "Учётные данные",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_go-park-mail-ru_2026_1_WHITECROWSOFT_internal_auth_dto.SignInUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Аутентификация успешна, в Set-Cookie возвращается JWT",
                        "schema": {
                            "$ref": "#/definitions/github_com_go-park-mail-ru_2026_1_WHITECROWSOFT_internal_auth_dto.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректные входные данные",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Неверный логин или пароль",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "405": {
                        "description": "Пустое тело запроса",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/signup": {
            "post": {
                "description": "Создаёт нового пользователя и устанавливает JWT-cookie сессии.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Регистрация нового пользователя",
                "parameters": [
                    {
                        "description": "Данные регистрации",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_go-park-mail-ru_2026_1_WHITECROWSOFT_internal_auth_dto.SignUpUser"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пользователь создан, в Set-Cookie возвращается JWT",
                        "schema": {
                            "$ref": "#/definitions/github_com_go-park-mail-ru_2026_1_WHITECROWSOFT_internal_auth_dto.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректные входные данные",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "405": {
                        "description": "Пустое тело запроса",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "Пользователь уже существует",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "github_com_go-park-mail-ru_2026_1_WHITECROWSOFT_internal_auth_dto.SignInUser": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_go-park-mail-ru_2026_1_WHITECROWSOFT_internal_auth_dto.SignUpUser": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "github_com_go-park-mail-ru_2026_1_WHITECROWSOFT_internal_auth_dto.UserResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "JWT-токен сессии, выставляется сервером после /signup или /signin.",
            "type": "apiKey",
            "name": "token",
            "in": "cookie"
        },
        "CsrfToken": {
            "description": "CSRF-токен, полученный из /csrf-token. Требуется для POST/PUT/DELETE.",
            "type": "apiKey",
            "name": "X-CSRF-Token",
            "in": "header"
        }
    },
...
}